+++ requirements/openvino.txt
@@ -4,3 +4,4 @@ networkx~=2.5
 nncf>=2.1.0
 onnx>=1.10.1
 openvino-dev>=2022.3.0
+openvino-model-api>=0.1.5
