+++ src/anomalib/deploy/inferencers/__init__.py
@@ -3,8 +3,6 @@
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .base_inferencer import Inferencer
-from .openvino_inferencer import OpenVINOInferencer
 from .torch_inferencer import TorchInferencer
 
-__all__ = ["Inferencer", "OpenVINOInferencer", "TorchInferencer"]
+__all__ = ["TorchInferencer"]
